[onboarding] renamed SelectableLink to Option
