--- src/components/onboarding/__stories__/selectable_link.story.tsx
@@ -1,11 +1,0 @@
-import { storiesOf } from "@storybook/react"
-import * as React from "react"
-import SelectableLink from "../selectable_link"
-
-storiesOf("Onboarding", module).add("SelectableLink", () => {
-  return (
-    <div style={{ width: "400px" }}>
-      <SelectableLink href="#" text="Buy Art & Design" onSelect={(select) => select} />
-    </div>
-  )
-})
