+++ src/components/onboarding/steps/collector_intent/__stories__/option.story.tsx
@@ -1,0 +1,11 @@
+import { storiesOf } from "@storybook/react"
+import * as React from "react"
+import Option from "../option"
+
+storiesOf("Onboarding", module).add("Option", () => {
+  return (
+    <div style={{ width: "400px" }}>
+      <Option href="#" text="Buy Art &amp; Design" onSelect={(select) => select} />
+    </div>
+  )
+})
